+++ modules/processings/provider.py
@@ -73,6 +73,7 @@ from .copyFeaturesIntoRoadElement import CopyFeaturesIntoRoadElement
 from .layoutItemsToCanvas import LayoutItemsToCanvasAlgorithm
 from .generateOmMapJson import GenerateOmMapJson
 from .batchGenerateOmMapJson import BatchGenerateOmMapJson
+from .simplifyLandUseNames import SimplifyCadastralNames
 
 
 class Provider(QgsProcessingProvider):
@@ -130,6 +131,7 @@ class Provider(QgsProcessingProvider):
         self.addAlgorithm(BuildingGeneralizationAlgorithm())
         self.addAlgorithm(GenerateOmMapJson())
         self.addAlgorithm(BatchGenerateOmMapJson())
+        self.addAlgorithm(SimplifyCadastralNames())
 
     def load(self):
         ProcessingConfig.settingIcons[self.name()] = self.icon()